+++ rs/registry/canister-client/src/data_provider.rs
@@ -1,3 +1,4 @@
+use crate::stable_memory::{StableMemoryBorrower, StorableRegistryKey, StorableRegistryValue};
 use candid::Principal;
 use ic_interfaces_registry::{
     RegistryDataProvider, RegistryTransportRecord, ZERO_REGISTRY_VERSION,
@@ -7,76 +8,12 @@ use ic_registry_transport::pb::v1::RegistryDelta;
 use ic_registry_transport::{
     deserialize_get_changes_since_response, serialize_get_changes_since_request,
 };
-use ic_stable_structures::memory_manager::VirtualMemory;
-use ic_stable_structures::storable::Bound;
-use ic_stable_structures::{DefaultMemoryImpl, StableBTreeMap, Storable};
 use ic_types::registry::RegistryDataProviderError;
 use ic_types::RegistryVersion;
 use itertools::Itertools;
-use std::borrow::Cow;
 use std::cmp::Ordering;
 use std::collections::HashSet;
 use std::marker::PhantomData;
-
-pub struct StorableRegistryValue(Option<Vec<u8>>);
-
-impl Storable for StorableRegistryValue {
-    fn to_bytes(&self) -> Cow<'_, [u8]> {
-        self.0.to_bytes()
-    }
-
-    fn from_bytes(bytes: Cow<'_, [u8]>) -> Self {
-        Self(Option::from_bytes(bytes))
-    }
-
-    const BOUND: Bound = Bound::Unbounded;
-}
-
-#[derive(Clone, Ord, PartialOrd, Eq, PartialEq, Default)]
-pub struct StorableRegistryKey {
-    version: u64,
-    key: String,
-}
-
-// This value is set as 2 times the max key size present in the registry
-const MAX_REGISTRY_KEY_SIZE: u32 = 200;
-
-impl Storable for StorableRegistryKey {
-    fn to_bytes(&self) -> Cow<[u8]> {
-        let mut storable_key = vec![];
-        let version_b = self.version.to_be_bytes().to_vec();
-        let key_b = self.key.as_bytes().to_vec();
-
-        storable_key.extend_from_slice(&version_b);
-        storable_key.extend_from_slice(&key_b);
-
-        Cow::Owned(storable_key)
-    }
-
-    fn from_bytes(bytes: Cow<[u8]>) -> Self {
-        let bytes = bytes.as_ref();
-        let (version_bytes, key_bytes) = bytes.split_at(8);
-        let version = u64::from_be_bytes(version_bytes.try_into().expect("Invalid version bytes"));
-        let key = String::from_utf8(key_bytes.to_vec()).expect("Invalid UTF-8 in key");
-
-        Self { version, key }
-    }
-    const BOUND: Bound = Bound::Bounded {
-        max_size: MAX_REGISTRY_KEY_SIZE + size_of::<u64>() as u32,
-        is_fixed_size: false,
-    };
-}
-
-type VM = VirtualMemory<DefaultMemoryImpl>;
-
-pub trait StableMemoryBorrower: Send + Sync {
-    fn with_borrow<R>(
-        f: impl FnOnce(&StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VM>) -> R,
-    ) -> R;
-    fn with_borrow_mut<R>(
-        f: impl FnOnce(&mut StableBTreeMap<StorableRegistryKey, StorableRegistryValue, VM>) -> R,
-    ) -> R;
-}
 
 /// This registry data provider is designed to work with the `ic-registry-canister-client`
 /// in canisters, enabling the retrieval and storage of a registry copy in stable memory.
@@ -231,12 +168,13 @@ impl<S: StableMemoryBorrower> RegistryDataProvider for CanisterDataProvider<S> {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::data_provider::{
-        CanisterDataProvider, StableMemoryBorrower, StorableRegistryKey, StorableRegistryValue,
-    };
+    use crate::data_provider::CanisterDataProvider;
+    use crate::stable_memory::{StableMemoryBorrower, StorableRegistryKey, StorableRegistryValue};
     use ic_registry_transport::pb::v1::RegistryDelta;
     use ic_stable_structures::memory_manager::MemoryId;
     use ic_stable_structures::memory_manager::MemoryManager;
+    use ic_stable_structures::memory_manager::VirtualMemory;
+    use ic_stable_structures::DefaultMemoryImpl;
     use ic_stable_structures::StableBTreeMap;
     use std::cell::RefCell;
     use std::collections::HashSet;
